Exercise_classification_and_Analysis/LLM/requirements.txt
flask 
python-dotenv 
google-generativeai 
flask-session
mediapipe
numpy
opencv-python
Werkzeug
